Slider: Move slider into widgets folder
