+++ tests/unit/slider/methods.js
@@ -1,6 +1,6 @@
 define( [
 	"jquery",
-	"ui/slider"
+	"ui/widgets/slider"
 ], function( $ ) {
 
 module( "slider: methods" );
